Signed-off-by: Gage Black <gtblack4@gmail.com>
Changed the funtions that display posts and comments.
Now when you select a submission or comment to reply to, instead of looping back through the comments until it reaches the number you chose, it stores all the post ID's in an array and posts the comments to the ID of the selection.

Renamed the file as well.